requirements.txt
asgiref==3.3.1
dj-database-url==0.5.0
Django==3.1.7
django-crispy-forms==1.11.1
django-heroku==0.3.1
gunicorn==20.0.4
Pillow==8.1.0
psycopg2==2.8.6
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
whitenoise==5.2.0
